refactor(voice/health): split contract.py into subpackage [AP-16, T1.1]
Closes anti-pattern #16 hygiene for contract.py (1982 LOC, mixed
responsibilities — 7 enums + 30 dataclasses + 1 Protocol + 8
platform-validation tables). Pure refactor: zero behaviour change.

Layout per godfile-splits mission Part 3 / T01:

  voice/health/contract/
    __init__.py       — re-exports every public-by-history symbol (49)
    _diagnosis.py     — Diagnosis enum (leaf — no internal deps)
    _combo.py         — Combo, CandidateEndpoint, CandidateSource,
                        CascadeResult, ComboEntry, OverrideEntry,
                        ProbeMode, LoadReport, ComboStoreStats +
                        ALLOWED_* validation tables + _platform_key
    _probe_result.py  — ProbeResult, ProbeHistoryEntry, RemediationHint,
                        AudioSubsystemFingerprint, CaptureTaskProto
    _bypass.py        — IntegrityVerdict / Result, BypassVerdict /
                        Context / Outcome, watchdog event lifecycle
                        (HotplugEvent, PowerEvent, AudioServiceEvent,
                        WatchdogState + *Kind enums)
    _eligibility.py   — Eligibility, all Mixer* types
                        (snapshot stack + L2.5 KB + sanity result),
                        HardwareContext

Cross-submodule dependencies are forward-only — no cycles:

    _diagnosis  → (none)
    _combo      → _diagnosis (TYPE_CHECKING), _probe_result (TYPE_CHECKING)
    _probe_result → _combo, _diagnosis (both TYPE_CHECKING)
    _bypass     → _probe_result.CaptureTaskProto (TYPE_CHECKING)
    _eligibility → _diagnosis, _probe_result.RemediationHint (both
                   TYPE_CHECKING)

Forward references are safe because every sub-file uses
`from __future__ import annotations` (PEP 563); annotations are
strings until inspected.

Spec deviations (one):

  * The spec listed `MixerControlSnapshot`, `MixerCardSnapshot`,
    `MixerApplySnapshot` under `_bypass.py`. Re-reading the original
    file, those are Linux-mixer snapshot types under the "ALSA mixer
    snapshots (Phase 3)" section, NOT bypass-strategy types. The "all
    Mixer* types" rule for `_eligibility.py` is the better fit:
    `MixerSanityResult.rollback_snapshot: MixerApplySnapshot | None`
    keeps the dependency intra-file rather than forcing a TYPE_CHECKING
    import across sub-files. Moved to `_eligibility.py`.

  * Did NOT re-export `_platform_key` / `_allowed_host_apis_for` from
    the package `__init__.py`: those were never in the original
    `__all__` and Wave 3 audit + grep confirms zero external imports.
    They remain importable via `sovyx.voice.health.contract._combo`
    if any code ever needs them.

Backwards compatibility: every `from sovyx.voice.health.contract import X`
in the codebase keeps working unchanged via the package `__init__.py`
re-exports. Wave 3 recon confirmed 0 test patches on
`patch("sovyx.voice.health.contract.X", ...)` — none to migrate.

Validation:
  * 49-name smoke import: OK
  * ruff check + format clean
  * mypy strict: 408 source files, 0 issues
  * bandit clean
  * pytest full suite: 11768 passed, 25 skipped, 57 deselected
    (identical to pre-split baseline)

Refs: docs-internal/missions/MISSION-voice-godfile-splits-v0.24.1.md
      Part 3 / T01
      docs-internal/missions/MISSION-voice-final-skype-grade-2026.md
      Phase 1 / T1.1
      CLAUDE.md anti-pattern #16
