--- src/sovyx/voice/health/contract.py
@@ -1,1982 +1,0 @@
-"""L0 — the contract every VCHL layer speaks.
-
-This module defines the vocabulary used across the Voice Capture Health
-Lifecycle: enums for diagnoses + probe modes, dataclasses for combos +
-probe results + persisted entries + cascade outcomes, plus the small set
-of platform-aware sanity validators that gate object creation.
-
-The contract is intentionally validation-strict at construction time so a
-malformed :class:`Combo` (e.g. 192 channels, sample rate 12345) cannot
-propagate beyond the boundary that built it. Persisted JSON entries
-re-validate on load via :class:`~sovyx.voice.health.combo_store.ComboStore`
-so a corrupted on-disk file cannot poison the runtime either.
-"""
-
-from __future__ import annotations
-
-import sys
-from dataclasses import dataclass, field
-from enum import StrEnum
-from typing import TYPE_CHECKING, Literal, Protocol, TypeAlias, runtime_checkable
-
-if TYPE_CHECKING:
-    from collections.abc import Awaitable, Callable, Mapping
-    from datetime import datetime
-    from pathlib import Path
-
-    import numpy as np
-    import numpy.typing as npt
-
-    from sovyx.voice._capture_task import (
-        AlsaHwDirectRestartResult,
-        ExclusiveRestartResult,
-        SessionManagerRestartResult,
-        SharedRestartResult,
-    )
-    from sovyx.voice.device_enum import DeviceEntry, DeviceKind
-
-
-# ── Enums ────────────────────────────────────────────────────────────────
-
-
-class Diagnosis(StrEnum):
-    """Outcome of a single probe.
-
-    Order matches the descending-confidence triage in
-    :func:`~sovyx.voice.health.probe.probe`. The cascade treats only
-    :attr:`HEALTHY` as a winning combo; every other value triggers
-    fallthrough or remediation.
-    """
-
-    HEALTHY = "healthy"
-    MUTED = "muted"
-    NO_SIGNAL = "no_signal"
-    LOW_SIGNAL = "low_signal"
-    FORMAT_MISMATCH = "format_mismatch"
-    APO_DEGRADED = "apo_degraded"
-    VAD_INSENSITIVE = "vad_insensitive"
-    DRIVER_ERROR = "driver_error"
-    DEVICE_BUSY = "device_busy"
-    HOT_UNPLUGGED = "hot_unplugged"
-    SELF_FEEDBACK = "self_feedback"
-    PERMISSION_DENIED = "permission_denied"
-    # Kernel-side IAudioClient invalidated state: device enumerates as
-    # healthy (PnP status=OK, ConfigManager=0) but every host API returns
-    # paInvalidDevice (-9996) on stream open because the IMMDevice's
-    # internal ``IAudioClient::Initialize`` path is stuck. Triggered by
-    # USB resource timeouts (LiveKernelEvent 0x1cc), driver hot-swaps,
-    # or mid-stream PnP churn. No user-mode cure exists — sovyx must
-    # quarantine the endpoint and fail-over to the next available
-    # capture device. Cure is physical: replug or reboot.
-    KERNEL_INVALIDATED = "kernel_invalidated"
-    # L2.5 mixer sanity diagnoses — emitted by
-    # :func:`sovyx.voice.health._mixer_sanity.check_and_maybe_heal`. Distinct
-    # from ``APO_DEGRADED`` because the fix path is mixer-layer (not APO
-    # bypass) and the root cause is pre-PortAudio gain misconfiguration,
-    # not capture-side DSP. See ADR-voice-mixer-sanity-l2.5-bidirectional §2.
-    MIXER_ZEROED = "mixer_zeroed"
-    """Attenuation regime — factory mixer gain too low, KB match found.
-    Fix is applied in-place by L2.5 via :class:`MixerPresetSpec`.
-    """
-
-    MIXER_SATURATED = "mixer_saturated"
-    """Saturation regime — factory mixer gain + boosts clipping internally.
-    Previously a sub-case of ``APO_DEGRADED``; split out as first-class so
-    the bypass coordinator routes to the mixer reset path instead of APO
-    bypass strategies. Existing ``LinuxALSAMixerResetBypass`` eligibility
-    extends to this diagnosis.
-    """
-
-    MIXER_UNKNOWN_PATTERN = "mixer_unknown_pattern"
-    """Mixer state outside the healthy range but no KB profile matches.
-    L2.5 defers; cascade proceeds to the platform walk. Telemetry flags
-    the hardware for prioritisation in KB growth.
-    """
-
-    MIXER_CUSTOMIZED = "mixer_customized"
-    """User customization detected (6-signal heuristic score > 0.75).
-    L2.5 is a no-op; the user's intentional tuning is preserved per
-    invariant I4.
-    """
-
-    MIXER_CALIBRATION_NEEDED = "mixer_calibration_needed"
-    """No KB match AND no customization signal — first-run candidate
-    hardware that needs the setup wizard to produce a validated preset.
-    """
-
-    UNKNOWN = "unknown"
-
-
-class CandidateSource(StrEnum):
-    """Reason a :class:`CandidateEndpoint` is in the cascade's candidate set.
-
-    Populated by :func:`~sovyx.voice.health._candidate_builder.build_capture_candidates`.
-    Used for telemetry (``voice_cascade_probe_call.candidate_source``) and
-    for the UI banner that explains why the cascade picked a fallback.
-
-    Members:
-        USER_PREFERRED: The device the user pinned via ``mind.yaml``
-            (``input_device_name``) or the request body. Always rank 0
-            when present.
-        CANONICAL_SIBLING: Same :attr:`~sovyx.voice.device_enum.DeviceEntry.canonical_name`
-            as USER_PREFERRED but different ``host_api_name`` — legacy
-            "same physical device exposed by multiple host APIs" path.
-            Dominant on Windows (WASAPI / DirectSound / MME of one mic);
-            usually empty on Linux where PortAudio exposes only ALSA.
-        SESSION_MANAGER_VIRTUAL: PipeWire / PulseAudio / JACK virtual
-            PCM. Added on Linux only, and only when
-            ``USER_PREFERRED.kind == HARDWARE`` (we don't duplicate).
-        OS_DEFAULT: The ``default`` / ``sysdefault`` ALSA alias. Added
-            on Linux only, and only when not already present via
-            ``USER_PREFERRED``.
-        FALLBACK: Any remaining enumerated input device not already
-            claimed by one of the above sources. Tail of the list —
-            rarely reached; exists so a truly exotic setup (vendor
-            virtual PCM) still gets a shot before ``CaptureInoperativeError``.
-    """
-
-    USER_PREFERRED = "user_preferred"
-    CANONICAL_SIBLING = "canonical_sibling"
-    SESSION_MANAGER_VIRTUAL = "session_manager_virtual"
-    OS_DEFAULT = "os_default"
-    FALLBACK = "fallback"
-
-
-class ProbeMode(StrEnum):
-    """How a probe should validate the device.
-
-    * :attr:`COLD` — boot-time, no user assumed to be speaking. Validates
-      that the stream opens and PortAudio callbacks fire. Cannot detect
-      :attr:`Diagnosis.APO_DEGRADED` because a silent room produces the
-      same RMS as a destroyed signal.
-    * :attr:`WARM` — wizard / first-interaction / fix mode, user is
-      asked to speak. Runs the audio through SileroVAD to derive the
-      full diagnosis surface.
-    """
-
-    COLD = "cold"
-    WARM = "warm"
-
-
-class HotplugEventKind(StrEnum):
-    """Classification of an OS-level audio-device hot-plug notification.
-
-    The watchdog (:mod:`~sovyx.voice.health.watchdog`) translates every
-    platform-specific event (``WM_DEVICECHANGE`` on Windows,
-    ``udev`` on Linux, ``kAudioHardwarePropertyDevices`` on macOS) into
-    one of these kinds so downstream logic stays platform-agnostic.
-
-    :attr:`DEFAULT_DEVICE_CHANGED` is emitted by Sprint 2 Task #18 — the
-    constant lives here so the enum is complete from the start and the
-    watchdog's ``on_hotplug`` handler doesn't need a value-update when
-    the default-change listener lands.
-    """
-
-    DEVICE_ADDED = "device_added"
-    DEVICE_REMOVED = "device_removed"
-    DEFAULT_DEVICE_CHANGED = "default_device_changed"
-
-
-class PowerEventKind(StrEnum):
-    """OS-level power-management events observed by the watchdog.
-
-    ADR §4.4.4. Sleep invalidates PortAudio sessions and may re-enumerate
-    devices in a different order on resume; the pipeline must checkpoint
-    on suspend and re-cascade on resume.
-
-    Platform mapping:
-
-    * Windows — ``WM_POWERBROADCAST``: ``PBT_APMSUSPEND`` →
-      :attr:`SUSPEND`, ``PBT_APMRESUMEAUTOMATIC`` → :attr:`RESUME`.
-    * Linux — ``org.freedesktop.login1`` D-Bus ``PrepareForSleep``
-      (``True`` → suspend, ``False`` → resume). Landed in Sprint 4.
-    * macOS — ``IORegisterForSystemPower`` callbacks. Landed in Sprint 4.
-    """
-
-    SUSPEND = "suspend"
-    RESUME = "resume"
-
-
-class AudioServiceEventKind(StrEnum):
-    """Audio-subsystem service lifecycle events observed by the watchdog.
-
-    ADR §4.4.5. Windows ``audiosrv`` can die (driver crash, Windows
-    Update, user-initiated ``net stop``); when it restarts, PortAudio
-    streams that were open beforehand are permanently broken. The
-    watchdog reacts by stalling probes until :attr:`UP`, then triggers a
-    re-cascade.
-
-    Platform mapping:
-
-    * Windows — poll ``Get-Service audiosrv`` equivalent via ``sc query``
-      or ``pywin32``. :attr:`DOWN` is surfaced on transition Running→Stopped
-      (or on repeated PaHostError patterns that imply the service died).
-      :attr:`UP` fires once the service is Running again.
-    * Linux — Sprint 4 (``systemctl is-active pipewire.service`` +
-      ``pulseaudio.service`` when applicable).
-    * macOS — macOS ``coreaudiod`` is managed by launchd and generally
-      respawns; this listener is effectively a Noop on darwin.
-    """
-
-    DOWN = "down"
-    UP = "up"
-
-
-class WatchdogState(StrEnum):
-    """§4.4.1 lifecycle states exposed by :class:`VoiceCaptureWatchdog`.
-
-    * :attr:`IDLE` — baseline; no degradation observed. Hot-plug adds
-      are ignored, removes of the active endpoint still trigger a
-      re-cascade.
-    * :attr:`BACKOFF` — one or more warm re-probes are scheduled after
-      a call to :meth:`VoiceCaptureWatchdog.report_deafness`. The
-      endpoint transitions back to :attr:`IDLE` on the first HEALTHY
-      re-probe.
-    * :attr:`DEGRADED` — the backoff schedule exhausted without a
-      HEALTHY probe. The pipeline runs in push-to-talk-only mode until
-      the user reboots or a new viable device hot-plugs in.
-    """
-
-    IDLE = "idle"
-    BACKOFF = "backoff"
-    DEGRADED = "degraded"
-
-
-# ── Validation tables (immutable) ───────────────────────────────────────
-
-
-ALLOWED_SAMPLE_RATES: frozenset[int] = frozenset(
-    {8_000, 16_000, 22_050, 24_000, 32_000, 44_100, 48_000, 88_200, 96_000, 192_000},
-)
-"""Sample rates we accept in a :class:`Combo`. Wider than what we use in
-practice so cascade overrides can target unusual hardware."""
-
-
-ALLOWED_FORMATS: frozenset[str] = frozenset({"int16", "int24", "float32"})
-"""PortAudio sample formats we know how to drive end-to-end through the
-:class:`~sovyx.voice.FrameNormalizer` resampler."""
-
-
-_ALLOWED_CHANNELS_MIN = 1
-_ALLOWED_CHANNELS_MAX = 8
-_ALLOWED_FRAMES_PER_BUFFER_MIN = 64
-_ALLOWED_FRAMES_PER_BUFFER_MAX = 8_192
-
-
-ALLOWED_HOST_APIS_BY_PLATFORM: Mapping[str, frozenset[str]] = {
-    "win32": frozenset(
-        {
-            "WASAPI",
-            "Windows WASAPI",
-            "WDM-KS",
-            "Windows WDM-KS",
-            "DirectSound",
-            "Windows DirectSound",
-            "MME",
-        },
-    ),
-    "linux": frozenset({"ALSA", "PulseAudio", "PipeWire", "JACK"}),
-    "darwin": frozenset({"CoreAudio", "Core Audio"}),
-}
-"""Host APIs we accept on each platform.
-
-Both bare (``"WASAPI"``) and PortAudio-formatted (``"Windows WASAPI"``)
-labels are accepted so callers can use whichever the upstream API gave
-them without an extra normalization step.
-"""
-
-
-def _platform_key() -> str:
-    """Return the current platform key for :data:`ALLOWED_HOST_APIS_BY_PLATFORM`.
-
-    Splits the hairsplitting of ``sys.platform`` ("linux", "linux2",
-    "win32", "darwin", …) into the three buckets we care about and
-    falls back to ``"linux"`` for unknown POSIX-likes.
-    """
-    plat = sys.platform
-    if plat.startswith("win"):
-        return "win32"
-    if plat == "darwin":
-        return "darwin"
-    return "linux"
-
-
-def _allowed_host_apis_for(platform_key: str) -> frozenset[str]:
-    return ALLOWED_HOST_APIS_BY_PLATFORM.get(platform_key, frozenset())
-
-
-# ── Dataclasses ─────────────────────────────────────────────────────────
-
-
-@dataclass(frozen=True, slots=True)
-class Combo:
-    """The audio configuration tuple that opens a capture stream.
-
-    Validated at construction so an invalid tuple (out-of-range rate,
-    unknown host API for the current platform, channels = 0) raises
-    immediately instead of propagating into the cascade.
-
-    Args:
-        host_api: Host-API label (``"WASAPI"`` or PortAudio's
-            ``"Windows WASAPI"`` — both accepted). Validated against
-            :data:`ALLOWED_HOST_APIS_BY_PLATFORM`.
-        sample_rate: Device-side sample rate in Hz. Pipeline downstream
-            resamples to 16 kHz mono int16 regardless.
-        channels: Channel count requested from PortAudio. The
-            normalizer mixes down to mono.
-        sample_format: PortAudio sample format (``"int16"`` / ``"int24"``
-            / ``"float32"``).
-        exclusive: WASAPI exclusive mode flag (Windows only meaningful;
-            ignored on other platforms).
-        auto_convert: Let WASAPI resample / rechannel transparently
-            (Windows only meaningful).
-        frames_per_buffer: PortAudio buffer size in samples. ``480`` is
-            30 ms at 16 kHz and matches the Silero VAD window cadence.
-        platform_key: Optional override for the platform key used during
-            validation. Tests pass ``"win32"`` to validate Windows
-            cascades on a non-Windows host without monkey-patching
-            :mod:`sys`.
-
-    Raises:
-        ValueError: On any out-of-range field or platform-host_api
-            mismatch.
-    """
-
-    host_api: str
-    sample_rate: int
-    channels: int
-    sample_format: str
-    exclusive: bool
-    auto_convert: bool
-    frames_per_buffer: int
-    platform_key: str = ""  # empty = use current platform; set in tests
-
-    def __post_init__(self) -> None:
-        if not self.host_api:
-            msg = "host_api must be a non-empty string"
-            raise ValueError(msg)
-        plat = self.platform_key or _platform_key()
-        allowed = _allowed_host_apis_for(plat)
-        if allowed and self.host_api not in allowed:
-            msg = (
-                f"host_api={self.host_api!r} is not allowed on platform "
-                f"{plat!r}; expected one of {sorted(allowed)}"
-            )
-            raise ValueError(msg)
-        if self.sample_rate not in ALLOWED_SAMPLE_RATES:
-            msg = (
-                f"sample_rate={self.sample_rate} is not allowed; "
-                f"expected one of {sorted(ALLOWED_SAMPLE_RATES)}"
-            )
-            raise ValueError(msg)
-        if not _ALLOWED_CHANNELS_MIN <= self.channels <= _ALLOWED_CHANNELS_MAX:
-            msg = (
-                f"channels={self.channels} out of range "
-                f"[{_ALLOWED_CHANNELS_MIN}, {_ALLOWED_CHANNELS_MAX}]"
-            )
-            raise ValueError(msg)
-        if self.sample_format not in ALLOWED_FORMATS:
-            msg = (
-                f"sample_format={self.sample_format!r} not allowed; "
-                f"expected one of {sorted(ALLOWED_FORMATS)}"
-            )
-            raise ValueError(msg)
-        if not (
-            _ALLOWED_FRAMES_PER_BUFFER_MIN
-            <= self.frames_per_buffer
-            <= _ALLOWED_FRAMES_PER_BUFFER_MAX
-        ):
-            msg = (
-                f"frames_per_buffer={self.frames_per_buffer} out of range "
-                f"[{_ALLOWED_FRAMES_PER_BUFFER_MIN}, {_ALLOWED_FRAMES_PER_BUFFER_MAX}]"
-            )
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class CandidateEndpoint:
-    """One endpoint the cascade should consider when opening capture.
-
-    A :class:`CandidateEndpoint` is the cascade-land analogue of a
-    :class:`~sovyx.voice.device_enum.DeviceEntry`: it carries the
-    device identity *plus* the metadata the cascade needs to (a) order
-    attempts, (b) key the ComboStore, (c) drive quarantine lifecycle,
-    and (d) populate user-visible telemetry. Built by
-    :func:`~sovyx.voice.health._candidate_builder.build_capture_candidates`.
-
-    The distinction between :attr:`device_index` and
-    :attr:`~sovyx.voice.device_enum.DeviceEntry.index` is deliberate:
-    the cascade may produce multiple ``CandidateEndpoint``\\ s with
-    different ``device_index`` values derived from the same physical
-    device (e.g. ``hw:1,0`` + ``plughw:1,0`` on Linux), and each gets
-    its own :attr:`endpoint_guid` for independent fast-path / quarantine
-    state in :class:`~sovyx.voice.health.combo_store.ComboStore`.
-
-    Attributes:
-        device_index: PortAudio device index — ephemeral across boots,
-            but always valid for the lifetime of one enumeration pass.
-        host_api_name: PortAudio host-API name (``"ALSA"``, ``"Windows
-            WASAPI"``, ``"Core Audio"``, …). Used for telemetry and to
-            cross-check with the device's actual host API in the probe.
-        kind: Semantic classification
-            (:class:`~sovyx.voice.device_enum.DeviceKind`). Drives
-            Linux-specific candidate-builder ordering heuristics.
-        canonical_name: The normalised, host-API-agnostic device name —
-            equal to :attr:`DeviceEntry.canonical_name`. Used by the
-            cascade-side logging and ComboStore keys.
-        friendly_name: Human-readable device name, for UI banners and
-            doctor-subcommand output.
-        source: Why this candidate is in the set — see
-            :class:`CandidateSource`.
-        preference_rank: 0 = try first. Stable deterministic ordering
-            within one boot; **not** a persistent score. The
-            :class:`CascadeResult.winning_candidate`'s rank is what the
-            ComboStore fast-path hydrates on the next boot.
-        endpoint_guid: Stable identifier from
-            :func:`~sovyx.voice.health._factory_integration.derive_endpoint_guid`.
-            Each ``CandidateEndpoint`` gets its own — two candidates
-            that differ only by ``device_index`` but share
-            ``(canonical_name, host_api_name, platform)`` produce the
-            same guid by design (deterministic hash). Two candidates
-            with different ``canonical_name`` produce different guids
-            — which is exactly the desired ComboStore independence.
-    """
-
-    device_index: int
-    host_api_name: str
-    kind: DeviceKind
-    canonical_name: str
-    friendly_name: str
-    source: CandidateSource
-    preference_rank: int
-    endpoint_guid: str
-    default_samplerate: int = 0
-
-    def __post_init__(self) -> None:
-        if self.device_index < 0:
-            msg = f"device_index must be >= 0, got {self.device_index}"
-            raise ValueError(msg)
-        if not self.host_api_name:
-            msg = "host_api_name must be a non-empty string"
-            raise ValueError(msg)
-        if not self.endpoint_guid:
-            msg = "endpoint_guid must be a non-empty string"
-            raise ValueError(msg)
-        if self.preference_rank < 0:
-            msg = f"preference_rank must be >= 0, got {self.preference_rank}"
-            raise ValueError(msg)
-        if self.default_samplerate < 0:
-            msg = f"default_samplerate must be >= 0, got {self.default_samplerate}"
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class RemediationHint:
-    """Single-source-of-truth pointer to user-facing remediation copy.
-
-    The :attr:`code` is an i18n key; the frontend resolves it via
-    ``useTranslation()`` and the CLI resolves it server-side. Backend
-    code never inlines a translatable sentence.
-
-    Attributes:
-        code: Stable i18n key, e.g. ``"remediation.muted"``. Listed in
-            :mod:`sovyx.voice.health._remediation`.
-        severity: One of ``"info"`` / ``"warn"`` / ``"error"``.
-        cli_action: Optional CLI command that fixes the issue, e.g.
-            ``"sovyx doctor voice --fix"``. ``None`` when the
-            remediation is purely manual.
-    """
-
-    code: str
-    severity: str
-    cli_action: str | None = None
-
-    def __post_init__(self) -> None:
-        if not self.code:
-            msg = "remediation code must be non-empty"
-            raise ValueError(msg)
-        if self.severity not in {"info", "warn", "error"}:
-            msg = f"severity={self.severity!r} not in {{info, warn, error}}"
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class ProbeResult:
-    """Outcome of a single probe.
-
-    Args:
-        diagnosis: Triage label.
-        mode: Whether this was a cold or warm probe.
-        combo: The combo that was probed.
-        vad_max_prob: Max SileroVAD speech probability across the probe
-            window. Cold probes set this to ``None`` (VAD not run).
-        vad_mean_prob: Mean SileroVAD speech probability. ``None`` for
-            cold probes.
-        rms_db: RMS of the probe window in dBFS.
-        callbacks_fired: Number of PortAudio callback invocations
-            observed during the probe.
-        duration_ms: Actual probe duration (may be < requested when an
-            early-exit condition triggered).
-        error: PortAudio / OS exception text when the open failed.
-        remediation: Optional pointer to user-facing remediation copy.
-    """
-
-    diagnosis: Diagnosis
-    mode: ProbeMode
-    combo: Combo
-    vad_max_prob: float | None
-    vad_mean_prob: float | None
-    rms_db: float
-    callbacks_fired: int
-    duration_ms: int
-    error: str | None = None
-    remediation: RemediationHint | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class AudioSubsystemFingerprint:
-    """SHA256 snapshot of OS-level audio configuration.
-
-    Used to detect cumulative Windows updates / PulseAudio config
-    changes / CoreAudio HAL plugin churn that would otherwise slip past
-    a coarse ``platform.version()`` check. Only the SHA fields are
-    load-bearing; the timestamp is diagnostic.
-
-    Attributes:
-        windows_audio_endpoints_sha: SHA256 over the MMDevices subtree
-            on Windows (empty on other platforms).
-        windows_fxproperties_global_sha: SHA256 over every active
-            endpoint's FxProperties subtree (Windows only).
-        linux_pulseaudio_config_sha: SHA256 over PulseAudio /
-            PipeWire config files (Linux only).
-        macos_coreaudio_plugins_sha: SHA256 over the CoreAudio HAL
-            plugin list (macOS only).
-        computed_at: ISO-8601 UTC timestamp.
-    """
-
-    windows_audio_endpoints_sha: str = ""
-    windows_fxproperties_global_sha: str = ""
-    linux_pulseaudio_config_sha: str = ""
-    macos_coreaudio_plugins_sha: str = ""
-    computed_at: str = ""
-
-
-@dataclass(frozen=True, slots=True)
-class ProbeHistoryEntry:
-    """One element of :attr:`ComboEntry.probe_history` (ring buffer).
-
-    Bounded to the last 10 probes per endpoint so the on-disk file stays
-    small even after months of operation.
-    """
-
-    ts: str  # ISO-8601 UTC
-    mode: ProbeMode
-    diagnosis: Diagnosis
-    vad_max_prob: float | None
-    rms_db: float
-    duration_ms: int
-
-
-@dataclass(frozen=True, slots=True)
-class ComboEntry:
-    """One row of :class:`~sovyx.voice.health.combo_store.ComboStore`.
-
-    Pure data; persistence rules + invalidation gates live in the store.
-    """
-
-    endpoint_guid: str
-    device_friendly_name: str
-    device_interface_name: str
-    device_class: str
-    endpoint_fxproperties_sha: str
-    winning_combo: Combo
-    validated_at: str
-    validation_mode: ProbeMode
-    vad_max_prob_at_validation: float | None
-    vad_mean_prob_at_validation: float | None
-    rms_db_at_validation: float
-    probe_duration_ms: int
-    detected_apos_at_validation: tuple[str, ...]
-    cascade_attempts_before_success: int
-    boots_validated: int
-    last_boot_validated: str
-    last_boot_diagnosis: Diagnosis
-    probe_history: tuple[ProbeHistoryEntry, ...] = ()
-    pinned: bool = False
-    needs_revalidation: bool = False  # set by R6/R7/R8/R9/R10/R11 on load
-    # voice-linux-cascade-root-fix T11 / schema v3.
-    # :class:`~sovyx.voice.device_enum.DeviceKind` value as string.
-    # Back-compat default ``"unknown"`` preserves v2 entries as-is.
-    candidate_kind: str = "unknown"
-
-
-@dataclass(frozen=True, slots=True)
-class OverrideEntry:
-    """One row of :class:`~sovyx.voice.health.capture_overrides.CaptureOverrides`."""
-
-    endpoint_guid: str
-    device_friendly_name: str
-    pinned_combo: Combo
-    pinned_at: str
-    pinned_by: str  # "user" | "wizard" | "cli"
-    reason: str = ""
-
-
-@dataclass(frozen=True, slots=True)
-class HotplugEvent:
-    """Platform-agnostic view of one audio-device hot-plug notification.
-
-    Populated by :mod:`~sovyx.voice.health._hotplug_win` /
-    :mod:`~sovyx.voice.health._hotplug_linux` /
-    :mod:`~sovyx.voice.health._hotplug_mac`; consumed by
-    :class:`~sovyx.voice.health.watchdog.VoiceCaptureWatchdog`.
-
-    :attr:`endpoint_guid` mirrors the GUID the cascade uses as its
-    :class:`ComboStore` key. It is best-effort: some OS events report a
-    device path or an interface name only, in which case this field is
-    ``None`` and the watchdog compares :attr:`device_friendly_name`
-    against the active endpoint's friendly name instead. When both
-    fields are empty the watchdog treats the event as a generic
-    add/remove signal (no endpoint-scoped action).
-    """
-
-    kind: HotplugEventKind
-    endpoint_guid: str | None = None
-    device_friendly_name: str | None = None
-    device_interface_name: str | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class PowerEvent:
-    """Platform-agnostic view of one OS power-management notification.
-
-    ADR §4.4.4. Emitted by :class:`~sovyx.voice.health._power.PowerEventListener`
-    and consumed by :meth:`~sovyx.voice.health.watchdog.VoiceCaptureWatchdog.on_power_event`.
-    """
-
-    kind: PowerEventKind
-
-
-@dataclass(frozen=True, slots=True)
-class AudioServiceEvent:
-    """Platform-agnostic view of one audio-service lifecycle transition.
-
-    ADR §4.4.5. Emitted by
-    :class:`~sovyx.voice.health._audio_service.AudioServiceMonitor` and
-    consumed by :meth:`~sovyx.voice.health.watchdog.VoiceCaptureWatchdog.on_audio_service_event`.
-    """
-
-    kind: AudioServiceEventKind
-
-
-@dataclass(frozen=True, slots=True)
-class CascadeResult:
-    """Outcome of one :func:`~sovyx.voice.health.cascade.run_cascade` call.
-
-    Attributes:
-        winning_combo: The combo that produced :attr:`Diagnosis.HEALTHY`
-            (warm) or callbacks-OK (cold). ``None`` when the cascade
-            exhausted without a winner.
-        winning_candidate: The :class:`CandidateEndpoint` that yielded
-            the winner, with its original ``source`` and
-            ``preference_rank`` preserved. ``None`` on exhaustion,
-            quarantine, or when the cascade ran against a legacy
-            single-endpoint call-site (pre-T3 of voice-linux-cascade-
-            root-fix — kept for a window during the migration).
-        winning_probe: The :class:`ProbeResult` that earned the winning
-            combo. ``None`` on exhaustion.
-        attempts: Every probe that ran during the cascade, in order.
-            Always non-empty after a successful run; ``None`` when the
-            cascade short-circuited on a fast-path lookup that produced
-            no fresh probe.
-        attempts_count: Number of attempts before success (0 = ComboStore
-            fast path or pinned override hit). Independent of
-            :attr:`attempts` length so callers don't have to subtract 1
-            for the winning attempt.
-        budget_exhausted: ``True`` when the total wall-clock budget ran
-            out before a winner emerged.
-        source: Where the winning combo came from —
-            ``"pinned"`` / ``"store"`` / ``"cascade"`` / ``"none"`` /
-            ``"quarantined"``. ``"quarantined"`` means every candidate
-            was in the §4.4.7 kernel-invalidated quarantine and no probe
-            ran; callers should fail-over to the next viable endpoint.
-        endpoint_guid: GUID of the endpoint the cascade ran on (echoed
-            for caller convenience). With candidate-set cascade this
-            is the **winning** candidate's guid, or the first
-            candidate's guid on exhaustion (for log correlation).
-    """
-
-    endpoint_guid: str
-    winning_combo: Combo | None
-    winning_probe: ProbeResult | None
-    attempts: tuple[ProbeResult, ...]
-    attempts_count: int
-    budget_exhausted: bool
-    source: str  # "pinned" | "store" | "cascade" | "none" | "quarantined"
-    winning_candidate: CandidateEndpoint | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class LoadReport:
-    """Result of :meth:`~sovyx.voice.health.combo_store.ComboStore.load`.
-
-    Attributes:
-        rules_applied: Pairs of ``(rule_code, scope)`` where ``scope`` is
-            either an endpoint GUID or ``"<global>"``.
-        entries_loaded: Number of valid entries available after load.
-        entries_dropped: Number of entries discarded by sanity rules
-            (R12) or platform mismatch.
-        backup_used: ``True`` when the main file was unreadable and the
-            ``.bak`` file rescued the load.
-        archived_to: Optional path the corrupt file was moved to.
-    """
-
-    rules_applied: tuple[tuple[str, str], ...]
-    entries_loaded: int
-    entries_dropped: int
-    backup_used: bool
-    archived_to: Path | None
-
-
-@dataclass(slots=True)
-class ComboStoreStats:
-    """In-memory hit/miss counters since the last :meth:`load`.
-
-    Mutable on purpose so the store can update counters in place without
-    rebuilding a frozen object on every probe. Read-only from the
-    outside.
-    """
-
-    fast_path_hits: int = 0
-    fast_path_misses: int = 0
-    invalidations_by_reason: dict[str, int] = field(default_factory=dict)
-
-
-# ── Capture-integrity + bypass strategy (Phase 1) ───────────────────────
-#
-# These types generalise the Windows-only Voice-Clarity auto-bypass into an
-# OS-agnostic detection layer (:class:`IntegrityVerdict` / :class:`IntegrityResult`)
-# paired with a platform-specific strategy pattern
-# (:class:`BypassVerdict` / :class:`BypassContext` / :class:`BypassOutcome`
-# + :class:`PlatformBypassStrategy` Protocol in
-# :mod:`sovyx.voice.health.bypass._strategy`).
-#
-# See ``docs-internal/plans/voice-apo-os-agnostic-fix.md`` §2.3 for the
-# derivation of each field + threshold.
-
-
-class IntegrityVerdict(StrEnum):
-    """OS-agnostic classification of a live capture stream's signal quality.
-
-    Distinct from :class:`Diagnosis` because :class:`Diagnosis` mixes
-    stream-open outcomes (``DRIVER_ERROR``, ``DEVICE_BUSY``) with signal
-    analysis. :class:`IntegrityVerdict` is pure-signal: it is only ever
-    computed against a *live* capture stream whose ring buffer already
-    carries frames.
-
-    Members:
-        HEALTHY: RMS alive, VAD responsive, spectral envelope intact.
-        APO_DEGRADED: RMS alive but VAD dead AND spectral envelope
-            flattened — capture-side DSP (Windows Voice Clarity,
-            PulseAudio ``module-echo-cancel``, CoreAudio VPIO) is
-            destroying the signal before it reaches user space.
-        DRIVER_SILENT: RMS near zero / flat DC — the driver is open but
-            not delivering audio. Distinct from APO_DEGRADED because
-            the fix is different (reopen / re-enumerate vs APO bypass).
-        VAD_MUTE: VAD dead but spectrum intact and RMS in the noise
-            floor band — user is genuinely not speaking. Re-probe
-            later; not a fault.
-        INCONCLUSIVE: Probe aborted (timeout, teardown, insufficient
-            frames in ring buffer). Caller retries.
-    """
-
-    HEALTHY = "healthy"
-    APO_DEGRADED = "apo_degraded"
-    DRIVER_SILENT = "driver_silent"
-    VAD_MUTE = "vad_mute"
-    INCONCLUSIVE = "inconclusive"
-
-
-@dataclass(frozen=True, slots=True)
-class IntegrityResult:
-    """Outcome of one :class:`CaptureIntegrityProbe` run.
-
-    Populated by :meth:`~sovyx.voice.health.capture_integrity.CaptureIntegrityProbe.probe_warm`.
-    Every field is deterministic-from-input so tests can assert exact
-    values against synthesised signals.
-
-    Args:
-        verdict: Classification label.
-        endpoint_guid: The endpoint the probe ran against — echoed so
-            coordinators don't drift if the active endpoint rebinds
-            mid-probe.
-        rms_db: RMS of the probe window in dBFS. ``-inf`` normalised to
-            ``-120.0`` for finite-JSON friendliness.
-        vad_max_prob: Peak SileroVAD speech probability across the
-            window. In ``[0.0, 1.0]``.
-        spectral_flatness: Wiener entropy of the magnitude spectrum in
-            ``[0.0, 1.0]``. White noise ≈ 1.0, pure tone ≈ 0.0,
-            clean speech ≈ 0.10–0.15, APO-destroyed speech ≈ 0.28–0.35.
-        spectral_rolloff_hz: 85 %-energy roll-off frequency in Hz.
-            Voice-Clarity's low-pass pulls this below 4 kHz; clean
-            speech sits at 6–8 kHz.
-        duration_s: Actual probe duration (may be shorter than
-            requested when the buffer had fewer frames).
-        probed_at_utc: Timestamp of probe completion.
-        raw_frames: Number of int16 samples actually analysed.
-        detail: Optional diagnostic string — set on ``INCONCLUSIVE``
-            to explain *why* (e.g. ``"ring_buffer_underrun"``).
-    """
-
-    verdict: IntegrityVerdict
-    endpoint_guid: str
-    rms_db: float
-    vad_max_prob: float
-    spectral_flatness: float
-    spectral_rolloff_hz: float
-    duration_s: float
-    probed_at_utc: datetime
-    raw_frames: int
-    detail: str = ""
-
-
-class BypassVerdict(StrEnum):
-    """Outcome of one :meth:`PlatformBypassStrategy.apply` invocation.
-
-    A strategy reports exactly one verdict per ``apply``. The
-    coordinator reads the verdict + the post-apply :class:`IntegrityResult`
-    to decide whether to stop, advance to the next strategy, or revert.
-
-    Members:
-        APPLIED_HEALTHY: Strategy applied AND the subsequent integrity
-            re-probe returned HEALTHY. Terminal success.
-        APPLIED_STILL_DEAD: Strategy applied cleanly but the re-probe
-            still classifies the signal as APO_DEGRADED / DRIVER_SILENT.
-            Coordinator advances.
-        NOT_APPLICABLE: Strategy's :meth:`probe_eligibility` returned
-            :attr:`Eligibility.applicable=False` (e.g. Windows exclusive
-            mode disabled by policy, Linux ALSA hw node not exposed).
-        FAILED_TO_APPLY: ``apply`` itself raised or the underlying
-            restart verdict reported failure. Coordinator advances.
-        REVERTED: ``apply`` succeeded but ``revert`` was called
-            subsequently (strategy B proved strictly better than A).
-    """
-
-    APPLIED_HEALTHY = "applied_healthy"
-    APPLIED_STILL_DEAD = "applied_still_dead"
-    NOT_APPLICABLE = "not_applicable"
-    FAILED_TO_APPLY = "failed_to_apply"
-    REVERTED = "reverted"
-
-
-@dataclass(frozen=True, slots=True)
-class Eligibility:
-    """Feasibility report from :meth:`PlatformBypassStrategy.probe_eligibility`.
-
-    A strategy whose eligibility check returns ``applicable=False`` is
-    skipped by the coordinator without counting toward
-    ``bypass_strategy_max_attempts`` — a non-applicable strategy is not
-    an attempt.
-
-    Args:
-        applicable: ``True`` iff the strategy's preconditions are met
-            on the current endpoint + OS + tuning configuration.
-        reason: Machine-readable reason token. Stable across minor
-            versions so dashboards can key on it. Examples:
-            ``"exclusive_mode_disabled_by_policy"``,
-            ``"not_wasapi_endpoint"``, ``"alsa_hw_node_unavailable"``,
-            ``"not_implemented_phase_3_pipewire"``.
-        estimated_cost_ms: Informational forecast of how long the
-            subsequent ``apply`` is expected to take. Used by the
-            coordinator only for telemetry, never for sequencing.
-    """
-
-    applicable: bool
-    reason: str = ""
-    estimated_cost_ms: int = 0
-
-
-@runtime_checkable
-class CaptureTaskProto(Protocol):
-    """Minimum surface :class:`CaptureIntegrityCoordinator` needs from
-    :class:`sovyx.voice._capture_task.AudioCaptureTask`.
-
-    Defined here (not in ``_capture_task``) so bypass strategies depend
-    on the abstract Protocol, not the concrete class — keeps the
-    dependency direction clean and makes testing strategies trivial
-    (fake capture task is a plain object).
-    """
-
-    @property
-    def active_device_guid(self) -> str: ...
-
-    @property
-    def active_device_name(self) -> str: ...
-
-    @property
-    def active_device_index(self) -> int: ...
-
-    @property
-    def active_device_kind(self) -> str: ...
-
-    @property
-    def host_api_name(self) -> str | None: ...
-
-    async def request_exclusive_restart(self) -> ExclusiveRestartResult: ...
-
-    async def request_shared_restart(self) -> SharedRestartResult: ...
-
-    async def request_alsa_hw_direct_restart(self) -> AlsaHwDirectRestartResult: ...
-
-    async def request_session_manager_restart(
-        self,
-        target_device: DeviceEntry | None = None,
-    ) -> SessionManagerRestartResult: ...
-
-    async def tap_recent_frames(
-        self,
-        duration_s: float,
-    ) -> npt.NDArray[np.int16]: ...
-
-    def samples_written_mark(self) -> tuple[int, int]:
-        """Return an opaque ``(epoch, samples_written)`` pair.
-
-        The tuple is opaque: pass it unchanged to
-        :meth:`tap_frames_since_mark`; do not compare, do arithmetic
-        on, or log the individual components without acknowledging they
-        are implementation details of the ring-buffer state machine.
-
-        Implementations MUST produce the pair from a single atomic read
-        of the internal packed state so both ints are guaranteed to
-        correspond to the same state generation (same epoch). See
-        :mod:`sovyx.voice._capture_task` for the packed encoding
-        rationale.
-
-        v1.3 §4.2.2 rationale — the contract is ``tuple[int, int]``
-        rather than a single packed ``int`` so the pair survives every
-        JSON / Prometheus / structlog serialization boundary without
-        silently truncating (packed marks can exceed ``2**53``, the
-        JavaScript ``Number`` safe range). The packed representation
-        remains private to the capture task.
-
-        Synchronous. Safe to call from any async context.
-        """
-        ...
-
-    async def tap_frames_since_mark(
-        self,
-        mark: tuple[int, int],
-        min_samples: int,
-        max_wait_s: float,
-    ) -> npt.NDArray[np.int16]:
-        """Return frames written AFTER ``mark`` was captured.
-
-        Polls the ring-buffer state until at least ``min_samples`` new
-        frames have accumulated post-``mark`` or ``max_wait_s`` elapses,
-        then returns the tail slice via :meth:`tap_recent_frames`. The
-        returned array always corresponds to audio the capture task
-        delivered strictly after :meth:`samples_written_mark` was
-        called — the primary fix for the v0.21.2 probe-window
-        contamination bug (see dossier ``SVX-VOICE-LINUX-20260422``).
-
-        Ring-buffer reset handling: if the epoch bundled in ``mark``
-        differs from the current epoch, the buffer was reallocated
-        mid-session (e.g. a WASAPI exclusive restart zeroed the ring).
-        Implementations MUST treat every sample currently in the buffer
-        as post-mark and return whatever is available without waiting
-        for the absent pre-reset frame count.
-
-        Args:
-            mark: Opaque pair returned by
-                :meth:`samples_written_mark`. Pass unchanged; never
-                synthesize one.
-            min_samples: Target accumulation threshold — the method
-                returns as soon as this many new frames are available.
-            max_wait_s: Hard timeout. On expiry the method returns
-                whatever fresh frames exist (possibly empty) rather
-                than blocking the coordinator's bypass loop.
-
-        Returns:
-            Zero-copy-free slice of the ring buffer, at most
-            ``min_samples`` long. Empty array on timeout with no new
-            frames.
-        """
-        ...
-
-    def apply_mic_ducking_db(self, gain_db: float) -> None: ...
-
-
-@dataclass(frozen=True, slots=True)
-class BypassContext:
-    """Per-apply context handed to a :class:`PlatformBypassStrategy`.
-
-    Pure-data; no mutable references. The coordinator rebuilds this
-    object each attempt so strategies cannot race on shared state.
-
-    Args:
-        endpoint_guid: GUID of the endpoint the coordinator is
-            operating on.
-        endpoint_friendly_name: Human-readable mic name. Used for log
-            messages and dashboards, never for logic.
-        host_api_name: PortAudio host-API label (``"Windows WASAPI"``,
-            ``"ALSA"``, ``"CoreAudio"`` …). Strategies use this for
-            their eligibility checks.
-        platform_key: Normalised ``sys.platform`` bucket (``"win32"``
-            / ``"linux"`` / ``"darwin"``). Pre-computed so tests can
-            pin the bucket without monkey-patching :mod:`sys`.
-        capture_task: :class:`CaptureTaskProto` reference — the only
-            mutating edge a strategy has into the running pipeline.
-        probe_fn: Callable that re-runs the warm integrity probe
-            against the live capture stream. Strategies invoke it
-            after ``apply`` to validate effectiveness.
-    """
-
-    endpoint_guid: str
-    endpoint_friendly_name: str
-    host_api_name: str
-    platform_key: str
-    capture_task: CaptureTaskProto
-    probe_fn: Callable[[], Awaitable[IntegrityResult]]
-    current_device_index: int = -1
-    current_device_kind: str = "unknown"
-
-
-@dataclass(frozen=True, slots=True)
-class BypassOutcome:
-    """Record of one full strategy attempt — emitted as a log event.
-
-    Args:
-        strategy_name: The :attr:`PlatformBypassStrategy.name` of the
-            strategy that ran.
-        attempt_index: Position in the coordinator's per-session
-            iteration (0 = first strategy tried).
-        verdict: Overall outcome.
-        integrity_before: Integrity probe result captured immediately
-            before ``apply``.
-        integrity_after: Integrity probe result captured after the
-            post-apply settle window. ``None`` iff ``verdict`` is
-            :attr:`BypassVerdict.NOT_APPLICABLE` or
-            :attr:`BypassVerdict.FAILED_TO_APPLY` (no post-apply probe
-            was possible).
-        elapsed_ms: Wall-clock time from entering ``apply`` to emitting
-            the outcome.
-        detail: Free-form diagnostic string. Populated on failure paths
-            with the classified error.
-    """
-
-    strategy_name: str
-    attempt_index: int
-    verdict: BypassVerdict
-    integrity_before: IntegrityResult
-    integrity_after: IntegrityResult | None
-    elapsed_ms: float
-    detail: str = ""
-
-
-# ── Linux ALSA mixer snapshots (Phase 3) ────────────────────────────────
-#
-# These types describe the *state* of the ALSA analog mixer chain on one
-# card. They are produced by :mod:`sovyx.voice.health._linux_mixer_probe`
-# (read-only), consumed by :mod:`sovyx.voice.health._linux_mixer_apply`
-# (writes + snapshot-driven rollback), and serialised onto the dashboard
-# via ``GET /api/voice/linux-mixer-diagnostics``.
-#
-# See ``docs-internal/plans/linux-alsa-mixer-saturation-fix.md`` §2.3.2
-# for the derivation of each field + the classification rules.
-
-
-@dataclass(frozen=True, slots=True)
-class MixerControlSnapshot:
-    """State of one ``amixer`` simple control on a single ALSA card.
-
-    Produced by :func:`sovyx.voice.health._linux_mixer_probe.enumerate_alsa_mixer_snapshots`.
-    Immutable — apply/revert builds :class:`MixerApplySnapshot` from these,
-    never mutates them in place.
-
-    Args:
-        name: ``amixer`` simple-control name (``"Capture"``,
-            ``"Internal Mic Boost"``, …). Case-sensitive; used verbatim
-            in the subsequent ``amixer -c N set`` call.
-        min_raw: Lower bound of the control's raw integer range (usually
-            ``0`` but some codecs expose negative floors).
-        max_raw: Upper bound of the raw integer range. Reading
-            ``current_raw == max_raw`` on a boost-class control is the
-            canonical saturation signal.
-        current_raw: Current raw value. The probe reads the Front-Left
-            channel; controls with a left/right asymmetry the probe
-            picks up are surfaced via :attr:`asymmetric`.
-        current_db: Current gain in dB, when the control exposes a dB
-            mapping. ``None`` for enum controls or controls without a
-            dB tag.
-        max_db: DB value at :attr:`max_raw`, for aggregated-boost
-            accounting. ``None`` when :attr:`current_db` is ``None``.
-        is_boost_control: ``True`` when the control name matches one of
-            the boost / capture patterns recognised by the probe — the
-            set of controls that saturate ADCs when driven to max.
-        saturation_risk: ``True`` iff :attr:`is_boost_control` AND
-            ``current_raw / max_raw > linux_mixer_saturation_ratio_ceiling``.
-            The coordinator keys off this flag.
-        asymmetric: ``True`` when the Front-Left and Front-Right
-            readings differ. Surfaced for diagnostics only; apply sets
-            both channels to the same target value.
-    """
-
-    name: str
-    min_raw: int
-    max_raw: int
-    current_raw: int
-    current_db: float | None
-    max_db: float | None
-    is_boost_control: bool
-    saturation_risk: bool
-    asymmetric: bool = False
-
-
-@dataclass(frozen=True, slots=True)
-class MixerCardSnapshot:
-    """State of all gain-bearing controls on a single ALSA card.
-
-    Args:
-        card_index: ``/proc/asound/cards`` index. Used directly in
-            ``amixer -c <index>`` subsequent calls.
-        card_id: Short identifier from ``/proc/asound/cards`` (e.g.
-            ``"Generic_1"``).
-        card_longname: Full human-readable name (e.g.
-            ``"HDA Intel (Family 17h/19h HD Audio Controller)"``). Used
-            for endpoint matching against
-            :attr:`BypassContext.endpoint_friendly_name`.
-        controls: All controls observed on this card, in ``amixer``
-            enumeration order.
-        aggregated_boost_db: Sum of :attr:`MixerControlSnapshot.current_db`
-            across every boost-class control that exposes a dB mapping.
-            ``0.0`` when no boost control reports a dB value.
-        saturation_warning: ``True`` iff at least one control has
-            :attr:`MixerControlSnapshot.saturation_risk` OR
-            :attr:`aggregated_boost_db` exceeds
-            ``linux_mixer_aggregated_boost_db_ceiling``.
-    """
-
-    card_index: int
-    card_id: str
-    card_longname: str
-    controls: tuple[MixerControlSnapshot, ...]
-    aggregated_boost_db: float
-    saturation_warning: bool
-
-
-@dataclass(frozen=True, slots=True)
-class MixerApplySnapshot:
-    """Record of a completed ``amixer`` mutation — drives revert.
-
-    Produced by :func:`sovyx.voice.health._linux_mixer_apply.apply_mixer_reset`
-    on success. Consumed by :func:`restore_mixer_snapshot` to roll back
-    the strategy's mutation on coordinator teardown or next-strategy
-    advancement.
-
-    Args:
-        card_index: Card whose controls were mutated.
-        reverted_controls: ``(name, pre_apply_raw_value)`` pairs in the
-            order they were mutated. Revert walks this list in reverse
-            so the last mutation is undone first.
-        applied_controls: ``(name, post_apply_raw_value)`` pairs for
-            telemetry + the dashboard's confirmation UI. Never used for
-            revert — the revert-from-snapshot contract is "restore the
-            pre-apply state", not "set to a different target".
-        reverted_enum_controls: ``(name, pre_apply_enum_label)`` pairs
-            for enum-typed controls mutated during apply (chiefly
-            HDA ``Auto-Mute Mode``). Default ``()`` so callers that
-            only touch numeric controls need no new field. Paranoid-QA
-            R3 CRIT-1: added to fill the half-heal WAL coverage gap
-            — without it, a mid-apply crash between the numeric
-            mutations and the auto-mute write would be recovered with
-            numerics restored but Auto-Mute stuck in the applied
-            ``Disabled``/``Enabled`` state. Revert walks this list in
-            reverse AFTER (or before, per the LIFO-order fix)
-            ``reverted_controls``.
-    """
-
-    card_index: int
-    reverted_controls: tuple[tuple[str, int], ...]
-    applied_controls: tuple[tuple[str, int], ...]
-    reverted_enum_controls: tuple[tuple[str, str], ...] = ()
-
-
-# ── L2.5 mixer sanity (Phase F1) ────────────────────────────────────────
-#
-# These types implement the bidirectional mixer-sanity layer spec'd in
-# ADR-voice-mixer-sanity-l2.5-bidirectional and the V2 Master Plan Part D.
-# They sit between ComboStore's fast-path and the platform cascade walk
-# so attenuation (Incident B) and saturation (Incident A) regimes are
-# handled symmetrically, KB-driven, and with full rollback.
-#
-# Convention (mirrors the rest of this module):
-#   * User/config/YAML-boundary inputs validate in ``__post_init__``
-#     (``MixerKBProfile``, ``MixerPresetSpec``, ``ValidationGates``,
-#     ``FactorySignature``, ``VerificationRecord``,
-#     ``MixerPresetControl``, ``MixerPresetValueFraction``).
-#   * Pure computed outputs stay validation-free — the producer (the
-#     L2.5 orchestrator) is the sole author and tests are the safety net
-#     (``MixerSanityResult``, ``MixerValidationMetrics``,
-#     ``MixerPresetValueRaw``, ``MixerPresetValueDb``).
-
-
-class MixerControlRole(StrEnum):
-    """Canonical role of an ALSA ``amixer`` simple control.
-
-    Role-based discovery replaces driver-family-specific string matching
-    so HDA, SOF, USB-audio, and BT codecs are handled by the same KB
-    profile schema. The per-driver resolver (``_mixer_roles.py``, Phase
-    F1.B) maps raw control names to these roles; the KB (``_mixer_kb/``,
-    Phase F1.C) keys presets by role; ``_linux_mixer_apply.apply_mixer_preset``
-    (Phase F1.D) translates (role, value) back to (name, raw).
-
-    Members are spec'd in V2 Master Plan Appendix 1. ``UNKNOWN`` is the
-    sentinel for controls observed by the probe but not role-mapped —
-    they are surfaced for telemetry and ignored by preset apply.
-    """
-
-    # Capture path — gain + source
-    CAPTURE_MASTER = "capture_master"
-    INTERNAL_MIC_BOOST = "internal_mic_boost"
-    PREAMP_BOOST = "preamp_boost"
-    DIGITAL_CAPTURE = "digital_capture"
-    INPUT_SOURCE_SELECTOR = "input_source_selector"
-    # Muting
-    AUTO_MUTE = "auto_mute"
-    CAPTURE_SWITCH = "capture_switch"
-    # SOF topology specifics (Intel Tiger Lake / Meteor Lake / Lunar Lake)
-    PGA_MASTER = "pga_master"
-    PGA_DMIC = "pga_dmic"
-    # USB audio class
-    USB_MIC_MASTER = "usb_mic_master"
-    # Bluetooth
-    BT_HFP_GAIN = "bt_hfp_gain"
-    # Sentinel
-    UNKNOWN = "unknown"
-
-
-class MixerSanityDecision(StrEnum):
-    """Terminal verdict of one :func:`check_and_maybe_heal` invocation.
-
-    The cascade keys off the decision to decide whether to skip the
-    platform walk (``HEALED``) or proceed (``SKIPPED_*``, ``DEFERRED_*``,
-    ``ROLLED_BACK``, ``ERROR``). See V2 Master Plan Part D.1.2.
-    """
-
-    HEALED = "healed"
-    """Preset applied, post-apply validation gates passed, persisted via
-    ``alsactl store``. Cascade records the winning combo and returns
-    without walking the platform list.
-    """
-
-    ROLLED_BACK = "rolled_back"
-    """Preset applied but at least one validation gate failed. Original
-    mixer state restored via :func:`restore_mixer_snapshot`. Cascade
-    proceeds to the platform walk; dashboard surfaces a warning.
-    """
-
-    SKIPPED_HEALTHY = "skipped_healthy"
-    """Mixer probe reports state in the healthy range (no KB match
-    required). L2.5 is a no-op; cascade proceeds normally.
-    """
-
-    SKIPPED_CUSTOMIZED = "skipped_customized"
-    """User-customization heuristic score exceeded the skip threshold
-    (``> linux_mixer_user_customization_threshold_skip``). Invariant I4
-    — user customization is sacred. Dashboard tip emitted.
-    """
-
-    DEFERRED_NO_KB = "deferred_no_kb"
-    """State outside healthy range, no KB profile matched above
-    ``match_threshold``. Cascade proceeds; hardware flagged for KB
-    growth prioritisation.
-    """
-
-    DEFERRED_AMBIGUOUS = "deferred_ambiguous"
-    """Two or more KB profiles matched with scores within the
-    ambiguity window (0.05 by default). Dashboard offers a choice card;
-    L2.5 defers this boot.
-    """
-
-    DEFERRED_PLATFORM = "deferred_platform"
-    """Non-Linux platform (F1 scope is Linux-only). Windows + macOS
-    fall through to their existing cascade paths.
-    """
-
-    ERROR = "error"
-    """Probe, apply, or validation raised unexpectedly. Rollback
-    attempted if a mixer mutation was in flight. Error reason token
-    captured in :attr:`MixerSanityResult.error`.
-    """
-
-
-# Tagged-union members for preset values. Each variant is a frozen,
-# slotted dataclass so pattern-matching in ``apply_mixer_preset`` stays
-# explicit (``match value: case MixerPresetValueFraction(...)``).
-
-
-@dataclass(frozen=True, slots=True)
-class MixerPresetValueRaw:
-    """Preset value expressed as the codec's native raw integer.
-
-    Use when the KB author has validated the raw value against the
-    specific codec (usually via HIL ``amixer cget``). The apply layer
-    clamps to ``[min_raw, max_raw]`` of the live control.
-    """
-
-    raw: int
-
-
-@dataclass(frozen=True, slots=True)
-class MixerPresetValueFraction:
-    """Preset value expressed as a fraction of the control's raw range.
-
-    Portable across codecs with different raw ranges — ``fraction=1.0``
-    resolves to ``max_raw``, ``0.0`` to ``min_raw``. Validated at
-    construction to the closed unit interval ``[0, 1]``.
-    """
-
-    fraction: float
-
-    def __post_init__(self) -> None:
-        if not 0.0 <= self.fraction <= 1.0:
-            msg = (
-                f"fraction={self.fraction!r} must be in [0.0, 1.0]; "
-                "use MixerPresetValueRaw for out-of-range values"
-            )
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class MixerPresetValueDb:
-    """Preset value expressed in dB, translated to raw via the control's
-    ``amixer``-reported dB mapping.
-
-    Use when the KB author has a dB target from a forensic capture
-    (``Capture: 0.0 dB``). The apply layer looks up the closest raw
-    value whose ``dB`` column matches and clamps to the control's range.
-    """
-
-    db: float
-
-
-MixerPresetValue: TypeAlias = MixerPresetValueRaw | MixerPresetValueFraction | MixerPresetValueDb
-"""Tagged-union type for the value slot of a :class:`MixerPresetControl`.
-Apply layer pattern-matches on the concrete variant to translate to raw.
-"""
-
-
-_CHANNEL_POLICIES: frozenset[str] = frozenset({"all", "left_right_equal"})
-_AUTO_MUTE_MODES: frozenset[str] = frozenset({"disabled", "enabled", "leave"})
-_RUNTIME_PM_TARGETS: frozenset[str] = frozenset({"on", "auto", "leave"})
-_DRIVER_FAMILIES: frozenset[str] = frozenset({"hda", "sof", "usb-audio", "bt"})
-_FACTORY_REGIMES: frozenset[str] = frozenset(
-    {"attenuation", "saturation", "mixed", "either"},
-)
-_AUDIO_STACKS: frozenset[str] = frozenset({"pipewire", "pulseaudio", "alsa"})
-_SANITY_REGIMES: frozenset[str] = frozenset(
-    {"saturation", "attenuation", "mixed", "healthy", "unknown"},
-)
-
-
-@dataclass(frozen=True, slots=True)
-class MixerPresetControl:
-    """One control line in a :class:`MixerPresetSpec`.
-
-    Args:
-        role: Canonical role the preset targets. ``UNKNOWN`` is rejected —
-            an ``UNKNOWN`` role in a preset is always a KB-authoring bug
-            (either the YAML references a role that doesn't exist or the
-            resolver table lacks the mapping the preset assumes).
-        value: Desired value (raw/fraction/db variant).
-        channel_policy: How to drive stereo controls — ``"all"`` sets
-            every channel to the same target; ``"left_right_equal"``
-            enforces L/R symmetry by asserting they were equal before
-            apply (tests raise if asymmetric to protect user intent).
-
-    Raises:
-        ValueError: On ``role=UNKNOWN`` or unknown ``channel_policy``.
-    """
-
-    role: MixerControlRole
-    value: MixerPresetValue
-    channel_policy: Literal["all", "left_right_equal"] = "all"
-
-    def __post_init__(self) -> None:
-        if self.role is MixerControlRole.UNKNOWN:
-            msg = (
-                "role=MixerControlRole.UNKNOWN is not a valid preset target; "
-                "resolve the control via MixerControlRoleResolver before "
-                "building a MixerPresetControl"
-            )
-            raise ValueError(msg)
-        if self.channel_policy not in _CHANNEL_POLICIES:
-            msg = f"channel_policy={self.channel_policy!r} not in {sorted(_CHANNEL_POLICIES)}"
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class MixerPresetSpec:
-    """Immutable corrective preset, sourced from a KB profile.
-
-    Translated to actual ``amixer sset`` calls by
-    :func:`~sovyx.voice.health._linux_mixer_apply.apply_mixer_preset`
-    (Phase F1.D). Pure data — no side effects at construction.
-
-    Args:
-        controls: Non-empty sequence of per-role value assignments. Empty
-            presets are rejected because they would produce no apply
-            record and thus no rollback capability — a silent no-op is
-            worse than an explicit KB authoring error.
-        auto_mute_mode: Whether to toggle HDA Auto-Mute Mode. ``"leave"``
-            (default) preserves the current setting.
-        runtime_pm_target: Whether to request codec runtime_pm change.
-            Applied via the systemd oneshot (Phase F1.G), never from the
-            daemon directly (invariant I7).
-
-    Raises:
-        ValueError: On empty ``controls`` or unknown literal values.
-    """
-
-    controls: tuple[MixerPresetControl, ...]
-    auto_mute_mode: Literal["disabled", "enabled", "leave"] = "leave"
-    runtime_pm_target: Literal["on", "auto", "leave"] = "leave"
-
-    def __post_init__(self) -> None:
-        if not self.controls:
-            msg = "MixerPresetSpec.controls must be non-empty"
-            raise ValueError(msg)
-        if self.auto_mute_mode not in _AUTO_MUTE_MODES:
-            msg = f"auto_mute_mode={self.auto_mute_mode!r} not in {sorted(_AUTO_MUTE_MODES)}"
-            raise ValueError(msg)
-        if self.runtime_pm_target not in _RUNTIME_PM_TARGETS:
-            msg = (
-                f"runtime_pm_target={self.runtime_pm_target!r} not in "
-                f"{sorted(_RUNTIME_PM_TARGETS)}"
-            )
-            raise ValueError(msg)
-        # Each role may appear at most once — a preset with two entries
-        # for the same role is a KB bug; the apply layer would silently
-        # keep only the last write.
-        seen: set[MixerControlRole] = set()
-        for ctl in self.controls:
-            if ctl.role in seen:
-                msg = f"role={ctl.role.value!r} appears more than once in MixerPresetSpec.controls"
-                raise ValueError(msg)
-            seen.add(ctl.role)
-
-
-@dataclass(frozen=True, slots=True)
-class ValidationGates:
-    """Post-apply validation thresholds for one KB profile.
-
-    Every gate must pass for the apply to be considered successful;
-    any failure triggers rollback (invariant I5). Sourced from a KB
-    YAML ``validation:`` block; pure data — gate evaluation happens in
-    :func:`~sovyx.voice.health._mixer_sanity._validate_post_apply`.
-
-    Args:
-        rms_dbfs_range: Closed interval for RMS of the validation
-            window, dBFS. Both bounds must be ``<= 0`` (dBFS is
-            non-positive) and ``lo <= hi``.
-        peak_dbfs_max: Upper bound on peak dBFS (``<= 0``). Peak above
-            this indicates clipping after the preset.
-        snr_db_vocal_band_min: Minimum SNR in the 300–3400 Hz vocal
-            band, dB. Non-negative.
-        silero_prob_min: Minimum ``max_prob`` Silero VAD must emit over
-            the validation window. Closed unit interval.
-        wake_word_stage2_prob_min: Minimum OpenWakeWord stage-2 probe
-            probability. Closed unit interval.
-
-    Raises:
-        ValueError: On out-of-range values or inverted intervals.
-    """
-
-    rms_dbfs_range: tuple[float, float]
-    peak_dbfs_max: float
-    snr_db_vocal_band_min: float
-    silero_prob_min: float
-    wake_word_stage2_prob_min: float
-
-    def __post_init__(self) -> None:
-        lo, hi = self.rms_dbfs_range
-        if lo > hi:
-            msg = (
-                f"rms_dbfs_range={self.rms_dbfs_range!r} is inverted (lo > hi); expected lo <= hi"
-            )
-            raise ValueError(msg)
-        if lo > 0.0 or hi > 0.0:
-            msg = f"rms_dbfs_range={self.rms_dbfs_range!r} must be non-positive (dBFS <= 0)"
-            raise ValueError(msg)
-        if self.peak_dbfs_max > 0.0:
-            msg = f"peak_dbfs_max={self.peak_dbfs_max!r} must be non-positive (dBFS <= 0)"
-            raise ValueError(msg)
-        if self.snr_db_vocal_band_min < 0.0:
-            msg = f"snr_db_vocal_band_min={self.snr_db_vocal_band_min!r} must be non-negative"
-            raise ValueError(msg)
-        if not 0.0 <= self.silero_prob_min <= 1.0:
-            msg = f"silero_prob_min={self.silero_prob_min!r} must be in [0.0, 1.0]"
-            raise ValueError(msg)
-        if not 0.0 <= self.wake_word_stage2_prob_min <= 1.0:
-            msg = (
-                f"wake_word_stage2_prob_min={self.wake_word_stage2_prob_min!r} "
-                "must be in [0.0, 1.0]"
-            )
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class FactorySignature:
-    """Expected factory-bad reading for one control role.
-
-    Used by the classifier to disambiguate ``attenuation`` /
-    ``saturation`` / ``mixed`` regimes — a probe whose reading falls
-    inside one of the declared ranges is evidence the factory regime
-    matches the KB profile.
-
-    At least one of the three ``expected_*_range`` fields must be
-    non-``None``; an all-``None`` signature would match no reading and
-    contribute nothing to classification.
-
-    Args:
-        expected_raw_range: Closed inclusive range of raw integer
-            values, or ``None`` when the KB author keyed the signature
-            on fraction/dB instead.
-        expected_fraction_range: Closed inclusive range as fraction of
-            the control's raw span (``current_raw / max_raw``). Both
-            bounds in ``[0, 1]``; ``lo <= hi``.
-        expected_db_range: Closed inclusive dB range. Only meaningful
-            for controls whose ``amixer`` output includes a dB mapping.
-
-    Raises:
-        ValueError: On all-``None`` signature, inverted interval, or
-            fraction bounds outside the unit interval.
-    """
-
-    expected_raw_range: tuple[int, int] | None
-    expected_fraction_range: tuple[float, float] | None
-    expected_db_range: tuple[float, float] | None
-
-    def __post_init__(self) -> None:
-        if (
-            self.expected_raw_range is None
-            and self.expected_fraction_range is None
-            and self.expected_db_range is None
-        ):
-            msg = (
-                "FactorySignature requires at least one of "
-                "expected_raw_range / expected_fraction_range / "
-                "expected_db_range to be non-None"
-            )
-            raise ValueError(msg)
-        # Paranoid-QA HIGH #10: reject degenerate ranges (``lo == hi``)
-        # for fraction + db. The only range where a point-match is
-        # meaningful is ``expected_raw_range`` (``(0, 0)`` captures
-        # "boost muted" on HDA — a real pilot signature). Fraction +
-        # dB point-matches have near-zero probability of firing due
-        # to float precision and are almost certainly KB authoring
-        # bugs; reject loudly rather than ship a profile that never
-        # matches in production.
-        if self.expected_raw_range is not None:
-            lo_r, hi_r = self.expected_raw_range
-            if lo_r > hi_r:
-                msg = f"expected_raw_range={self.expected_raw_range!r} is inverted"
-                raise ValueError(msg)
-        if self.expected_fraction_range is not None:
-            lo_f, hi_f = self.expected_fraction_range
-            if lo_f > hi_f:
-                msg = f"expected_fraction_range={self.expected_fraction_range!r} is inverted"
-                raise ValueError(msg)
-            if not (0.0 <= lo_f <= 1.0 and 0.0 <= hi_f <= 1.0):
-                msg = (
-                    f"expected_fraction_range={self.expected_fraction_range!r} "
-                    "must lie within [0.0, 1.0]"
-                )
-                raise ValueError(msg)
-            if lo_f == hi_f:
-                msg = (
-                    f"expected_fraction_range={self.expected_fraction_range!r} "
-                    "is degenerate (lo == hi); use a band or widen the bounds"
-                )
-                raise ValueError(msg)
-        if self.expected_db_range is not None:
-            lo_d, hi_d = self.expected_db_range
-            if lo_d > hi_d:
-                msg = f"expected_db_range={self.expected_db_range!r} is inverted"
-                raise ValueError(msg)
-            if lo_d == hi_d:
-                msg = (
-                    f"expected_db_range={self.expected_db_range!r} is "
-                    "degenerate (lo == hi); use a band"
-                )
-                raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class VerificationRecord:
-    """One HIL verification attestation shipped with a KB profile.
-
-    Each profile's ``verified_on`` list must be non-empty before merge —
-    the reviewer pair records the exact system + kernel + distro the
-    profile was HIL-validated against. Empty fields are rejected so
-    ``grep``-ability of the provenance trail is preserved.
-
-    Args:
-        system_product: Exact product string (``dmidecode -s
-            system-product-name``).
-        codec_id: ``<vendor>:<device>`` hex pair from
-            ``/proc/asound/card*/codec#*`` or ``lspci``.
-        kernel: Exact ``uname -r`` output at verification time.
-        distro: Short distro name + version (e.g. ``"linuxmint-22.2"``).
-        verified_at: ISO-8601 date of verification.
-        verified_by: GitHub username or ``"sovyx-core"`` for first-party.
-
-    Raises:
-        ValueError: On any empty field.
-    """
-
-    system_product: str
-    codec_id: str
-    kernel: str
-    distro: str
-    verified_at: str
-    verified_by: str
-
-    def __post_init__(self) -> None:
-        for name, value in (
-            ("system_product", self.system_product),
-            ("codec_id", self.codec_id),
-            ("kernel", self.kernel),
-            ("distro", self.distro),
-            ("verified_at", self.verified_at),
-            ("verified_by", self.verified_by),
-        ):
-            if not value:
-                msg = f"VerificationRecord.{name} must be non-empty"
-                raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class MixerKBProfile:
-    """Validated knowledge-base entry for one hardware class.
-
-    Loaded from a YAML under ``_mixer_kb/profiles/*.yaml`` (Phase F1.C)
-    via pydantic v2 and re-constructed as this dataclass so downstream
-    consumers (classifier, apply layer, dashboard) see a single type.
-    The KB loader (not this dataclass) is responsible for Ed25519
-    signature verification when the shipped bundle is enabled.
-
-    Args:
-        profile_id: Snake-case unique identifier within the shipped KB.
-        profile_version: Monotonic integer starting at 1; bumps on any
-            semantic change to preset or factory signature.
-        schema_version: YAML schema version the profile was authored
-            against. Loader rejects mismatches.
-        codec_id_glob: ``fnmatch`` pattern over codec vendor:device
-            (e.g. ``"14F1:5045"``). Required — the primary match key.
-        driver_family: Which ALSA driver-family resolver applies.
-        system_vendor_glob: Optional ``fnmatch`` over dmidecode system
-            vendor.
-        system_product_glob: Optional ``fnmatch`` over dmidecode system
-            product.
-        distro_family: Optional coarse distro bucket (``ubuntu-like``
-            etc.). ``None`` means "any".
-        audio_stack: Optional constraint on the userspace audio stack.
-        kernel_major_minor_glob: Optional ``fnmatch`` over ``uname -r``.
-        match_threshold: Minimum weighted score required to select this
-            profile. Closed unit interval.
-        factory_regime: Which regime this profile's signature describes.
-        factory_signature: Per-role expected factory-bad readings.
-            ``Mapping`` is kept immutable-by-convention (loader builds
-            ``MappingProxyType`` wrappers).
-        recommended_preset: Corrective preset to apply on match.
-        validation_gates: Post-apply gates; all must pass.
-        verified_on: Non-empty tuple of HIL attestations.
-        contributed_by: Author GitHub username or ``"sovyx-core"``.
-
-    Raises:
-        ValueError: On any validation failure.
-    """
-
-    profile_id: str
-    profile_version: int
-    schema_version: int
-    codec_id_glob: str
-    driver_family: Literal["hda", "sof", "usb-audio", "bt"]
-    system_vendor_glob: str | None
-    system_product_glob: str | None
-    distro_family: str | None
-    audio_stack: Literal["pipewire", "pulseaudio", "alsa"] | None
-    kernel_major_minor_glob: str | None
-    match_threshold: float
-    factory_regime: Literal["attenuation", "saturation", "mixed", "either"]
-    factory_signature: Mapping[MixerControlRole, FactorySignature]
-    recommended_preset: MixerPresetSpec
-    validation_gates: ValidationGates
-    verified_on: tuple[VerificationRecord, ...]
-    contributed_by: str
-
-    def __post_init__(self) -> None:
-        if not self.profile_id:
-            msg = "MixerKBProfile.profile_id must be non-empty"
-            raise ValueError(msg)
-        if self.profile_version < 1:
-            msg = f"MixerKBProfile.profile_version={self.profile_version} must be >= 1"
-            raise ValueError(msg)
-        if self.schema_version < 1:
-            msg = f"MixerKBProfile.schema_version={self.schema_version} must be >= 1"
-            raise ValueError(msg)
-        if not self.codec_id_glob:
-            msg = "MixerKBProfile.codec_id_glob must be non-empty"
-            raise ValueError(msg)
-        if self.driver_family not in _DRIVER_FAMILIES:
-            msg = f"driver_family={self.driver_family!r} not in {sorted(_DRIVER_FAMILIES)}"
-            raise ValueError(msg)
-        if self.audio_stack is not None and self.audio_stack not in _AUDIO_STACKS:
-            msg = f"audio_stack={self.audio_stack!r} not in {sorted(_AUDIO_STACKS)}"
-            raise ValueError(msg)
-        if self.factory_regime not in _FACTORY_REGIMES:
-            msg = f"factory_regime={self.factory_regime!r} not in {sorted(_FACTORY_REGIMES)}"
-            raise ValueError(msg)
-        if not 0.0 <= self.match_threshold <= 1.0:
-            msg = f"match_threshold={self.match_threshold!r} must be in [0.0, 1.0]"
-            raise ValueError(msg)
-        if not self.factory_signature:
-            msg = "MixerKBProfile.factory_signature must be non-empty"
-            raise ValueError(msg)
-        if MixerControlRole.UNKNOWN in self.factory_signature:
-            msg = (
-                "MixerKBProfile.factory_signature must not contain "
-                "MixerControlRole.UNKNOWN — resolve roles before building "
-                "the profile"
-            )
-            raise ValueError(msg)
-        if not self.verified_on:
-            msg = (
-                "MixerKBProfile.verified_on must be non-empty — at least "
-                "one HIL attestation is required before merge"
-            )
-            raise ValueError(msg)
-        if not self.contributed_by:
-            msg = "MixerKBProfile.contributed_by must be non-empty"
-            raise ValueError(msg)
-
-
-@dataclass(frozen=True, slots=True)
-class MixerValidationMetrics:
-    """Signal-quality snapshot measured post-apply.
-
-    Pure output, populated by
-    :func:`~sovyx.voice.health._mixer_sanity._validate_post_apply`.
-    Compared field-by-field against a :class:`ValidationGates` to decide
-    accept vs rollback. All fields finite; ``-inf`` is normalised to
-    ``-120.0`` upstream for JSON friendliness (same convention as
-    :class:`IntegrityResult`).
-
-    Args:
-        rms_dbfs: RMS of the validation window, dBFS.
-        peak_dbfs: Peak magnitude, dBFS.
-        snr_db_vocal_band: SNR in the 300–3400 Hz band, dB.
-        silero_max_prob: Peak Silero VAD ``speech`` probability.
-        silero_mean_prob: Mean Silero VAD ``speech`` probability.
-        wake_word_stage2_prob: OpenWakeWord stage-2 probe result.
-        measurement_duration_ms: Wall-clock duration of the validation
-            window actually captured (may be shorter than requested if
-            the ring buffer drained early).
-    """
-
-    rms_dbfs: float
-    peak_dbfs: float
-    snr_db_vocal_band: float
-    silero_max_prob: float
-    silero_mean_prob: float
-    wake_word_stage2_prob: float
-    measurement_duration_ms: int
-
-
-@dataclass(frozen=True, slots=True)
-class MixerSanityResult:
-    """Terminal record of one :func:`check_and_maybe_heal` invocation.
-
-    Emitted by :mod:`sovyx.voice.health._mixer_sanity`. Consumed by
-    :func:`~sovyx.voice.health.cascade.run_cascade` to decide whether to
-    skip the platform cascade walk; by telemetry to feed the
-    ``mixer_sanity_outcome`` bucket; and by the dashboard to render the
-    Mixer Health card.
-
-    Args:
-        decision: Terminal verdict — see :class:`MixerSanityDecision`.
-        diagnosis_before: Classified state pre-intervention.
-        diagnosis_after: Classified state post-intervention. ``None``
-            when the decision short-circuited before a post-apply probe
-            (``SKIPPED_*``, ``DEFERRED_*``, ``ERROR``-before-probe).
-        regime: Categorical regime label. ``"healthy"`` when the probe
-            reported a healthy range and no KB match was required;
-            ``"unknown"`` when classification was inconclusive.
-        matched_kb_profile: ``profile_id`` of the KB profile that drove
-            the apply, or ``None`` when no profile matched.
-        kb_match_score: Weighted match score of the selected profile,
-            or 0.0 when no profile matched. In ``[0, 1]``.
-        user_customization_score: 6-signal heuristic result. In
-            ``[0, 1]``.
-        cards_probed: ``/proc/asound/cards`` indices seen during
-            probing.
-        controls_modified: Control names actually mutated during apply
-            (empty when decision was not ``HEALED`` or ``ROLLED_BACK``).
-        rollback_snapshot: The pre-apply snapshot retained for revert.
-            Set on ``HEALED`` (for long-tail rollback on later
-            invalidation) and on ``ROLLED_BACK`` (for telemetry).
-        probe_duration_ms: Wall-clock of the mixer probe step.
-        apply_duration_ms: Wall-clock of the apply step, or ``None``
-            when no apply ran.
-        validation_passed: Whether every :class:`ValidationGates` gate
-            was met. ``None`` when no validation ran.
-        validation_metrics: Captured metrics, or ``None`` when no
-            validation ran.
-        remediation: User-facing hint pointer for the dashboard/CLI;
-            populated for DEFERRED/SKIPPED_CUSTOMIZED/ERROR.
-        error: Machine-readable failure token from Appendix 6 of the
-            V2 plan (``"MIXER_SANITY_PROBE_FAILED"`` etc.).
-    """
-
-    decision: MixerSanityDecision
-    diagnosis_before: Diagnosis
-    diagnosis_after: Diagnosis | None
-    regime: Literal["saturation", "attenuation", "mixed", "healthy", "unknown"]
-    matched_kb_profile: str | None
-    kb_match_score: float
-    user_customization_score: float
-    cards_probed: tuple[int, ...]
-    controls_modified: tuple[str, ...]
-    rollback_snapshot: MixerApplySnapshot | None
-    probe_duration_ms: int
-    apply_duration_ms: int | None
-    validation_passed: bool | None
-    validation_metrics: MixerValidationMetrics | None
-    remediation: RemediationHint | None = None
-    error: str | None = None
-
-
-_HARDWARE_DRIVER_FAMILIES: frozenset[str] = frozenset(
-    {"hda", "sof", "usb-audio", "bt", "unknown"},
-)
-"""Driver families accepted by :class:`HardwareContext`.
-
-Superset of ``_DRIVER_FAMILIES`` (the KB-profile set) by the extra
-``"unknown"`` sentinel — KB profiles must target a specific family,
-but a detected hardware context may legitimately fail to determine
-one (e.g., obscure vendor driver on first boot before ``lspci`` /
-``/proc/asound/cards`` have populated).
-"""
-
-
-@dataclass(frozen=True, slots=True)
-class HardwareContext:
-    """Detected audio-hardware identity for L2.5 + KB matching.
-
-    Populated once per cascade pass by the Linux platform probe (or
-    the equivalent Windows/macOS stub in F3) and threaded through
-    three consumers:
-
-    * :class:`~sovyx.voice.health._mixer_roles.MixerControlRoleResolver`
-      — ``codec_id`` drives the per-codec override lookup;
-      ``driver_family`` drives the family-table lookup.
-    * ``MixerKBLookup.match`` (Phase F1.C) — every field is compared
-      ``fnmatch``-style against the corresponding glob on
-      :class:`MixerKBProfile`.
-    * ``_mixer_sanity._detect_user_customization`` (Phase F1.E) —
-      reads ``audio_stack`` / ``distro`` to locate the right config
-      paths for PipeWire / PulseAudio / WirePlumber.
-
-    Every field except ``driver_family`` is optional because detection
-    can partially fail — a missing ``codec_id`` should degrade the KB
-    match score, not abort the cascade.
-
-    Args:
-        driver_family: ALSA driver family. ``"unknown"`` is valid and
-            means "probe could not determine"; resolver then skips
-            the family table and relies on substring fallback.
-        codec_id: ``<vendor>:<device>`` hex pair read from
-            ``/proc/asound/card*/codec#*`` (e.g. ``"14F1:5045"``).
-            Case-normalisation is the producer's responsibility — the
-            resolver and KB loader both expect the exact string
-            ``/proc/asound`` reports.
-        system_vendor: ``dmidecode -s system-manufacturer`` output
-            (e.g. ``"Sony Group Corporation"``).
-        system_product: ``dmidecode -s system-product-name`` output
-            (e.g. ``"VJFE69F11X-B0221H"``).
-        distro: Short distro + version label (e.g.
-            ``"linuxmint-22.2"``). Used for config path resolution,
-            not for matching alone.
-        audio_stack: Active userspace audio stack, detected via
-            ``wpctl status`` / ``pactl info`` / ``/proc/asound/pcm``.
-            ``None`` when no userspace daemon is running.
-        kernel: Exact ``uname -r`` output (e.g. ``"6.14.0-37-generic"``).
-            Consumed as an ``fnmatch`` target against
-            :attr:`MixerKBProfile.kernel_major_minor_glob`.
-
-    Raises:
-        ValueError: On unknown ``driver_family`` or ``audio_stack``.
-    """
-
-    driver_family: Literal["hda", "sof", "usb-audio", "bt", "unknown"]
-    codec_id: str | None = None
-    system_vendor: str | None = None
-    system_product: str | None = None
-    distro: str | None = None
-    audio_stack: Literal["pipewire", "pulseaudio", "alsa"] | None = None
-    kernel: str | None = None
-
-    def __post_init__(self) -> None:
-        if self.driver_family not in _HARDWARE_DRIVER_FAMILIES:
-            msg = (
-                f"driver_family={self.driver_family!r} not in {sorted(_HARDWARE_DRIVER_FAMILIES)}"
-            )
-            raise ValueError(msg)
-        if self.audio_stack is not None and self.audio_stack not in _AUDIO_STACKS:
-            msg = f"audio_stack={self.audio_stack!r} not in {sorted(_AUDIO_STACKS)}"
-            raise ValueError(msg)
-
-
-__all__ = [
-    "ALLOWED_FORMATS",
-    "ALLOWED_HOST_APIS_BY_PLATFORM",
-    "ALLOWED_SAMPLE_RATES",
-    "AudioSubsystemFingerprint",
-    "BypassContext",
-    "BypassOutcome",
-    "BypassVerdict",
-    "CaptureTaskProto",
-    "CascadeResult",
-    "Combo",
-    "ComboEntry",
-    "ComboStoreStats",
-    "Diagnosis",
-    "Eligibility",
-    "FactorySignature",
-    "HardwareContext",
-    "IntegrityResult",
-    "IntegrityVerdict",
-    "LoadReport",
-    "MixerApplySnapshot",
-    "MixerCardSnapshot",
-    "MixerControlRole",
-    "MixerControlSnapshot",
-    "MixerKBProfile",
-    "MixerPresetControl",
-    "MixerPresetSpec",
-    "MixerPresetValue",
-    "MixerPresetValueDb",
-    "MixerPresetValueFraction",
-    "MixerPresetValueRaw",
-    "MixerSanityDecision",
-    "MixerSanityResult",
-    "MixerValidationMetrics",
-    "OverrideEntry",
-    "ProbeHistoryEntry",
-    "ProbeMode",
-    "ProbeResult",
-    "RemediationHint",
-    "ValidationGates",
-    "VerificationRecord",
-]
